renaming function names
