+++ pyzinga/model_eval_tools.py
@@ -1,7 +1,7 @@
 """                     Import libraries.                       """
 import lightgbm as lgbm
 import pandas as pd
-import utils.mlflow_utils as mlfu
+import pyzinga.mlflow_tools as mlfu
 from sklearn.metrics import (
     accuracy_score, f1_score, precision_score, recall_score, roc_auc_score, average_precision_score, log_loss, auc, 
     confusion_matrix, classification_report,
